refactor: 주문 상품 및 좋아요 타입 파일 마이페이지 types 폴더로 이동
